testing detalha
