+++ src/components/Exibir/TrabalhoDetalhado.css
@@ -1,0 +1,136 @@
+.cartaoDetalhado-container {
+    /* border: 2px solid brown;  */
+    display: flex;
+    flex-direction: column;
+    justify-content:flex-start;
+    align-items: center;
+    justify-items: center;
+    /* margin-bottom: 20px; */
+    /* margin: 10px ; */
+    margin: 30px ;
+    /* flex-wrap: wrap; */
+    width: 400px;
+    /* max-width: 100px; */
+    /* min-width: 200px; */
+    /* margin: 30px; */
+    /* height: 410px; */
+    background-color: #efffef;
+    /* padding: 10px; */
+    /* padding-top: 20px; */
+    border-radius: 10px;
+    box-shadow: 7px 7px 30px #777;
+}
+  
+.cartao-botao > button {
+    /* margin-left: 5px; */
+    /* margin-bottom: 2px; */
+    /* background-color:#eee; */
+    /* border: 1px solid rgb(84, 111, 111);  */
+    /* border-radius: 10px; */
+    outline: none;
+    /* padding: 5px 12px; */
+    display: block;
+    margin: 0 auto;
+    /* align-self: center;
+    justify-self: center; */
+    text-align: center;
+
+    }
+
+.cartao-foto {
+    /* margin-top: 20px; */
+    width: 100%;
+    /* width: 290px; */
+    /* min-width: 100px; */
+    height: 200px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    box-sizing: border-box;
+    /* border: 1px solid black; */
+    /* width: 200px; */
+    /* max-height: 120px; */
+    /* border: 1px solid blue; */
+    /* border-radius: 50px; */
+}
+
+img {
+    width: 100%;
+    /* width: 100%; */
+    height: 100%;
+    box-sizing: border-box;
+    background-size:cover;
+    background-repeat: no-repeat;
+    border-radius: 10px 10px 0 0;
+
+}
+
+.cartao-dados {
+    display: flex;
+    width: 400px;
+    flex-direction: column;
+    justify-content:space-between;
+    align-items:stretch;
+    align-content:stretch;
+    justify-items:stretch;
+    /* padding: 10px; */
+    font-size:large;
+    
+}
+
+
+.cartao-nome-e-valor {
+    display: flex;
+    flex-direction: row;
+    justify-content:space-between;
+    align-items:space-between;
+    justify-items:space-between;
+    padding: 10px;
+}
+
+
+
+
+.cartao-so-descricao {
+    padding: 10px ;
+    font-size: medium;
+    /* height: 80px; */
+}
+
+.cartao-metodo-e-prazo {
+    display: flex;
+    flex-direction: row;
+    justify-content:space-between;
+    align-items:space-between;
+    justify-items:space-between;
+    padding: 10px;
+    font-size:small;
+    color: grey;
+    padding-bottom: 10px;
+}
+
+.cartao-metodo {
+    /* background-color: blue; */
+}
+.cartao-prazo {
+    /* background-color: blue; */
+}
+
+
+
+.cartao-nome {
+    /* background-color: blue; */
+   }
+   .cartao-valor {
+       /* background-color: red; */
+       /* padding: 20px; */
+   }
+   
+
+button {
+    display: block;
+    margin: 0 auto;
+    margin-top: 30px; 
+    margin-bottom: 30px;
+    width: 250px;
+}
